selfCollision fixes
